% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_job_util.R
\name{get_utilization}
\alias{get_utilization}
\title{get_utilization - calculate resource utilization}
\usage{
get_utilization(sacct0, total_proc, dt = 60L)
}
\arguments{
\item{sacct0}{data frame with jobs execution history}

\item{total_proc}{total number of cores in the resource}

\item{dt}{time aggregation size, seconds. \emph{Default}: 60L}
}
\value{
data frame with resource utilization over time
}
\description{
Calculate resource utilization
}
\examples{

}
