% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_job_util.R
\name{read_simstat_backfill}
\alias{read_simstat_backfill}
\title{read_simstat_backfill - read backfill execution stats}
\usage{
read_simstat_backfill(filename)
}
\arguments{
\item{filename}{name of file to read}
}
\value{
data frame with resource utilization over time
}
\description{
Cread stats of backfill execution during simulated Slurm Run
}
\examples{

}
