% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_job_util.R
\name{sim_job}
\alias{sim_job}
\title{sim_job - Generates R-list with Resource Requested by Job}
\usage{
sim_job(job_id = NA, submit = "2001-01-01 00:00:00", wclimit = 0L,
  duration = 0L, tasks = 0L, tasks_per_node = 0L, username = "user1",
  qosname = "normal", partition = "normal", account = "account1",
  req_mem = NA, req_mem_per_cpu = 0L, features = "", gres = "",
  shared = 1L, cpus_per_task = 1L, dependency = "", cancelled = NA, ...)
}
\arguments{
\item{job_id}{Job ID. If it is not set then will use increamented \code{last_job_id values}. \emph{\emph{Default}}: NA}

\item{submit}{Date-Time to sumbit job. \emph{Default}: "2001-01-01 00:00:00"}

\item{wclimit}{Requested time in minutes. \emph{Default}: 0L}

\item{duration}{Simulated actual job walltime in seconds. \emph{Default}: 0L}

\item{tasks}{Requested number of tasks. \emph{Default}: 0L}

\item{tasks_per_node}{Requested number of tasks per node. \emph{Default}: 0L}

\item{username}{User name. \emph{Default}: "user1"}

\item{qosname}{QoS. \emph{Default}: "normal"}

\item{partition}{Partition. \emph{Default}: "normal"}

\item{account}{Account. \emph{Default}: "account1"}

\item{req_mem}{Requested memory in MiB. \emph{Default}: NA}

\item{req_mem_per_cpu}{Memory per what? 1L - memory per CPU, 0L - memory per whole job. \emph{Default}: 0L}

\item{features}{Requested features, a.k.a. constrains. \emph{Default}: ""}

\item{gres}{General Resources. \emph{Default}: "". Example: "gpu:2"}

\item{shared}{Is job run in shared mode. \emph{Default}: 1L}

\item{cpus_per_task}{CPUs per tast. \emph{Default}: 1L}

\item{dependency}{Dependencies on other jobs. \emph{Default}: "". Example: "afterok:1002:1003"}

\item{cancelled}{Date-Time to cancel job, if NA job would not be cancelled. \emph{Default}: NA}

\item{...}{Other arguments will be appended to returning list}
}
\value{
List with job's requested resources
}
\description{
Generates batch job request details as a list with named members.
All arguments have \emph{Default} vaalues. Several outputs from this function
can be combined together to data.frame to be fead to \code{\link{write_trace}}
}
\examples{
# Generate small job trace file
# job trace as list of lists
trace <- list(
    sim_job(
        job_id=1001,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        features="IB&CPU-M"
    ),sim_job(
        job_id=1002,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        dependency="afterok:1001"
    ),sim_job(
        job_id=1003,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        gres = "gpu:2"
    )
)
# convert list of lists to data.frame
trace <- do.call(rbind, lapply(trace,data.frame))
# write job trace
write_trace("test.trace",trace)

}
