% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_job_util.R
\name{write_trace}
\alias{write_trace}
\title{Write Trace Jobs File for Simulation}
\usage{
write_trace(trace_filename, trace)
}
\arguments{
\item{trace_filename}{filename for written trace job file}

\item{trace}{data.frame with requsted resource}
}
\description{
Write Trace Jobs File for Simulation.
}
\examples{
# Generate small job trace file
# job trace as list of lists
trace <- list(
    sim_job(
        job_id=1001,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        features="IB&CPU-M"
    ),sim_job(
        job_id=1002,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        dependency="afterok:1001"
    ),sim_job(
        job_id=1003,
        submit="2016-10-01 00:01:00",
        wclimit=300L,
        duration=600L,
        tasks=12L,
        tasks_per_node=12L,
        gres = "gpu:2"
    )
)
# convert list of lists to data.frame
trace <- do.call(rbind, lapply(trace,data.frame))
# write job trace
write_trace("test.trace",trace)

}
\seealso{
\code{\link{sim_job}}
}
